remove input, migrate style source input
