+++ apps/builder/app/builder/features/style-panel/style-source/style-source-input.tsx
@@ -19,8 +19,6 @@ import {
   ComboboxRoot,
   ComboboxAnchor,
   ComboboxContent,
-  DeprecatedTextFieldInput,
-  useDeprecatedTextFieldFocus,
   useCombobox,
   type CSS,
   ComboboxLabel,
@@ -33,6 +31,7 @@ import {
   styled,
   Flex,
   ComboboxScrollArea,
+  InputField,
 } from "@webstudio-is/design-system";
 import { CheckMarkIcon, DotIcon, LocalStyleIcon } from "@webstudio-is/icons";
 import {
@@ -43,6 +42,8 @@ import {
   type ForwardRefRenderFunction,
   type RefObject,
   type ReactNode,
+  useRef,
+  useCallback,
 } from "react";
 import { mergeRefs } from "@react-aria/utils";
 import { type ComponentState, stateCategories } from "@webstudio-is/react-sdk";
@@ -51,6 +52,7 @@ import { useSortable } from "./use-sortable";
 import { matchSorter } from "match-sorter";
 import { StyleSourceBadge } from "./style-source-badge";
 import { humanizeString } from "~/shared/string-utils";
+import { useFocusWithin } from "@react-aria/interactions";
 
 type IntermediateItem = {
   id: string;
@@ -84,7 +86,7 @@ const TextFieldContainer = styled("div", {
 });
 
 type TextFieldBaseWrapperProps<Item extends IntermediateItem> = Omit<
-  ComponentProps<"input">,
+  ComponentProps<typeof InputField>,
   "value"
 > &
   Pick<ComponentProps<typeof TextFieldContainer>, "css"> & {
@@ -113,7 +115,6 @@ const TextFieldBase: ForwardRefRenderFunction<
     onFocus,
     onBlur,
     onClick,
-    type,
     onKeyDown,
     label,
     value,
@@ -127,17 +128,25 @@ const TextFieldBase: ForwardRefRenderFunction<
     states,
     ...textFieldProps
   } = props;
-  const [internalInputRef, focusProps] = useDeprecatedTextFieldFocus({
-    onFocus,
-    onBlur,
-  });
   const { sortableRefCallback, dragItemId, placementIndicator } = useSortable({
     items: value,
     onSort,
   });
+  const internalInputRef = useRef<HTMLInputElement>(null);
+
+  const { focusWithinProps } = useFocusWithin({
+    onFocusWithin: onFocus,
+    onBlurWithin: onBlur,
+  });
+
+  const onClickCapture = useCallback(() => {
+    internalInputRef.current?.focus();
+  }, [internalInputRef]);
+
   return (
     <TextFieldContainer
-      {...focusProps}
+      {...focusWithinProps}
+      onClickCapture={onClickCapture}
       ref={mergeRefs(forwardedRef, containerRef ?? null, sortableRefCallback)}
       css={css}
       style={
@@ -147,8 +156,9 @@ const TextFieldBase: ForwardRefRenderFunction<
     >
       {/* We want input to be the first element in DOM so it receives the focus first */}
       {editingItemId === undefined && (
-        <DeprecatedTextFieldInput
+        <InputField
           {...textFieldProps}
+          variant="chromeless"
           css={{
             color: theme.colors.hiContrast,
             fontVariantNumeric: "tabular-nums",
@@ -156,11 +166,17 @@ const TextFieldBase: ForwardRefRenderFunction<
             fontSize: theme.deprecatedFontSize[3],
             lineHeight: 1,
             order: 1,
+            border: "none",
+            flex: 1,
+            "&:focus-within, &:hover": {
+              outline: "none",
+            },
           }}
+          size="1"
           value={label}
-          type={type}
           onClick={onClick}
-          ref={mergeRefs(internalInputRef, inputRef ?? null)}
+          ref={inputRef}
+          inputRef={internalInputRef}
           spellCheck={false}
           aria-label="New Style Source Input"
         />
