+++ packages/design-system/src/index.ts
@@ -88,5 +88,4 @@ export * from "./components/kbd";
 export * from "./components/__DEPRECATED__/text2";
 export { DeprecatedIconButton } from "./components/__DEPRECATED__/icon-button";
 export * from "./components/__DEPRECATED__/popover";
-export * from "./components/__DEPRECATED__/text-field";
 export * from "./components/__DEPRECATED__/list";
