+++ packages/design-system/src/stitches.config.ts
@@ -1,12 +1,9 @@
 import { createStitches } from "@stitches/react";
 import {
   blue,
-  blueA,
   gray,
   green,
-  greenA,
   red,
-  redA,
   slate,
   slateA,
   yellow,
@@ -67,11 +64,8 @@ const { styled, css, getCssText, globalCss, keyframes, config, reset } =
         slate7: slate.slate7,
         slate6: slate.slate6,
         slate8: slate.slate8,
-        slate9: slate.slate9,
         slate11: slate.slate11,
         slate12: slate.slate12,
-        slateA3: slateA.slateA3,
-        slateA4: slateA.slateA4,
         slateA7: slateA.slateA7,
         slateA8: slateA.slateA8,
         gray2: gray.gray2,
@@ -79,26 +73,14 @@ const { styled, css, getCssText, globalCss, keyframes, config, reset } =
         blue9: blue.blue9,
         blue10: blue.blue10,
         blue11: blue.blue11,
-        blueA3: blueA.blueA3,
-        blueA4: blueA.blueA4,
         red4: red.red4,
-        red8: red.red8,
         red9: red.red9,
         red11: red.red11,
         red10: red.red10,
-        redA3: redA.redA3,
-        redA4: redA.redA4,
-        redA8: redA.redA8,
-        green2: green.green2,
         green3: green.green3,
         green4: green.green4,
-        green7: green.green7,
-        green8: green.green8,
         green9: green.green9,
         green11: green.green11,
-        greenA3: greenA.greenA3,
-        greenA4: greenA.greenA4,
-        greenA8: greenA.greenA8,
         yellow9: yellow.yellow9,
         yellow10: yellow.yellow10,
         mint5: mint.mint5,
